+++ gameplay/src/BoundingSphere.cpp
@@ -38,7 +38,7 @@ bool BoundingSphere::intersects(const BoundingSphere& sphere) const
     float vy = sphere.center.y - center.y;
     float vz = sphere.center.z - center.z;
 
-    return sqrtf(vx * vx + vy * vy + vz * vz) <= (radius + sphere.radius);
+    return sqrt(vx * vx + vy * vy + vz * vz) <= (radius + sphere.radius);
 }
 
 bool BoundingSphere::intersects(const BoundingBox& box) const
@@ -86,7 +86,7 @@ bool BoundingSphere::intersects(const BoundingBox& box) const
     cpY -= center.y;
     cpZ -= center.z;
 
-    return sqrtf(cpX * cpX + cpY * cpY + cpZ * cpZ) <= radius;
+    return sqrt(cpX * cpX + cpY * cpY + cpZ * cpZ) <= radius;
 }
 
 bool BoundingSphere::intersects(const Frustum& frustum) const
@@ -144,7 +144,7 @@ float BoundingSphere::intersects(const Ray& ray) const
     else
     {
         // The intersection is at the smaller positive root.
-        float sqrtDisc = sqrtf(discriminant);
+        float sqrtDisc = sqrt(discriminant);
         float t0 = (-B - sqrtDisc) * 0.5f;
         float t1 = (-B + sqrtDisc) * 0.5f;
         return (t0 > 0.0f && t0 < t1) ? t0 : t1;
@@ -162,7 +162,7 @@ void BoundingSphere::merge(const BoundingSphere& sphere)
     float vx = center.x - sphere.center.x;
     float vy = center.y - sphere.center.y;
     float vz = center.z - sphere.center.z;
-    float d = sqrtf(vx * vx + vy * vy + vz * vz);
+    float d = sqrt(vx * vx + vy * vy + vz * vz);
 
     // If one sphere is contained inside the other, set to the larger sphere.
     if (d <= (sphere.radius - radius))
@@ -231,7 +231,7 @@ void BoundingSphere::merge(const BoundingBox& box)
     v1x = center.x - fx;
     v1y = center.y - fy;
     v1z = center.z - fz;
-    float distance = sqrtf(v1x * v1x + v1y * v1y + v1z * v1z);
+    float distance = sqrt(v1x * v1x + v1y * v1y + v1z * v1z);
 
     // If the box is inside the sphere, we are done.
     if (distance <= radius)
@@ -297,7 +297,7 @@ void BoundingSphere::transform(const Matrix& matrix)
 
 float BoundingSphere::distance(const BoundingSphere& sphere, const Vector3& point)
 {
-    return sqrtf((point.x - sphere.center.x) * (point.x - sphere.center.x) +
+    return sqrt((point.x - sphere.center.x) * (point.x - sphere.center.x) +
                  (point.y - sphere.center.y) * (point.y - sphere.center.x) +
                  (point.z - sphere.center.z) * (point.z - sphere.center.x));
 }
